requirements.txt
pydantic
openai
pandas
numpy
google-generativeai
python-dotenv
google-genai
beautifulsoup4
selenium
Flask
markdown
flask-sqlalchemy 
flask-login 
bcrypt
pyodbc
psycopg2
sqlalchemy
mysqlclient